style: move query to upper module
